move nilearn.plotting.edge_detect to nilearn.plotting.displays
