+++ nilearn/plotting/displays/edge_detect.py
